// packages/nocodb-sdk/ee/lib/Api.ts
export type BoolType = number | boolean | null;
// ...
export interface NormalColumnRequestType {
  /** Is this column auto-incremented? */
  ai?: BoolType;
  /** Is this column auto-updated datetime field? */
  au?: BoolType;
  /** Column Comment */
  cc?: StringOrNullType;
  /** Column Default Value */
  cdf?: StringOrNullType;
  /** Column Name */
  column_name: string;
  /** Model for StringOrNull */
  csn?: StringOrNullType;
  /** Data Type */
  dt?: string;
  /** Data Type Extra */
  dtx?: StringOrNullType;
  /** Data Type Extra Precision */
  dtxp?: string | number | null;
  /** Data Type Extra Scale */
  dtxs?: StringOrNullType | number;
  /** Numeric Precision */
  np?: StringOrNullType | number;
  /** Numeric Scale */
  ns?: StringOrNullType | number;
  /** Is this column a primary key? */
  pk?: BoolType;
  /** Is this column a primary value? */
  pv?: BoolType;
  /** Is this column required? */
  rqd?: BoolType;
  /** Column Title */
  title?: string;
  /** UI Data Type */
  uidt?:
    | 'Attachment'
    | 'AutoNumber'
    | 'Barcode'
    | 'Button'
    | 'Checkbox'
    | 'Collaborator'
    | 'Count'
    | 'CreateTime'
    | 'Currency'
    | 'Date'
    | 'DateTime'
    | 'Decimal'
    | 'Duration'
    | 'Email'
    | 'Formula'
    | 'ForeignKey'
    | 'GeoData'
    | 'Geometry'
    | 'ID'
    | 'JSON'
    | 'LastModifiedTime'
    | 'LongText'
    | 'LinkToAnotherRecord'
    | 'Lookup'
    | 'MultiSelect'
    | 'Number'
    | 'Percent'
    | 'PhoneNumber'
    | 'Rating'
    | 'Rollup'
    | 'SingleLineText'
    | 'SingleSelect'
    | 'SpecificDBType'
    | 'Time'
    | 'URL'
    | 'Year'
    | 'QrCode'
    | 'Links';
  /** Is this column unique? */
  un?: BoolType;
  /** Is this column unique? */
  unique?: BoolType;
}
// ...
export type StringOrNullType = string | null;